Keep replay preservation out of the chat monolith
Provider replay needs historical tool-input retention types, but growing message-prep crossed the repository's chat-module ceiling. Extract the stable type vocabulary into a focused module, preserve the existing message-prep exports, and tighten the ceiling to the smaller implementation.

Constraint: Existing imports from chat/message-prep must remain source-compatible.

Rejected: Raise the LOC ceiling | hides the structural regression and weakens the ratchet.

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep runtime compaction behavior in message-prep; this module owns only its shared type vocabulary.

Tested: chat ratchet; Deno check; provider replay/message preparation suite (71/71); lint:ci through all source policy gates, with only local Deno 2.7.12 vs pinned 2.7.7 docs-generation guard remaining.

Not-tested: Pinned-Deno docs generation locally; GitHub CI provides the pinned runtime.
